+++ source/extensions/isaacsim.sensors.physics.ui/isaacsim/sensors/physics/ui/tests/__init__.py
@@ -1,4 +1,4 @@
-# SPDX-FileCopyrightText: Copyright (c) 2018-2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-FileCopyrightText: Copyright (c) 2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -13,10 +13,5 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .file_validation import *
-from .image_capture import *
-from .image_comparison import *
-from .image_io import *
-from .menu_ui_test import *
-from .menu_utils import *
-from .timed_async_test import *
+from .test_context_menu import *
+from .test_menu import *